Use only one ModelAdmin for User model and use same regex for username everywhere
