+++ kobo/apps/accounts/models.py
@@ -1,4 +1,5 @@
 from allauth.account.signals import email_confirmed
+from django.contrib.auth.models import User
 from django.db import models
 from django.dispatch import receiver
